refactor: :recycle: refactor api folder
